refactor: refactor files and folders structure
